HP value bug fixed/ Port No. change to 8070
HP value was not shown correctly because client assume every player has
a HP value of 100, but what if it doesn't?(When the client respawn). So
passing in HP value from server, and change server code to send out HP
value.
Now the client need HP value for every newplayer as input parameter.
Port No. change to 8070 because for some reason, my computer has that
port occupied. (Or other wired problem.)